fix(sdk): keep a comparison row numeric, and bound the judge call
Review follow-ups on the n-way comparison work. Every finding was valid.

The comparison row key is a number again. `ComparisonOptions.index` accepted
`number | string`, but the server validates `evaluations[].index` as
`z.number()` and parses the body before dispatch, so a string row label failed
validation and took the whole batch down with it, the row's dataset entries
included. Narrowing matches Python, which already coerced and raised.

The evaluator fetch now carries a timeout. A judge that accepted the socket
and never answered held its slot in the concurrency window for the life of the
process, so one dead connection stalled a run rather than costing it one row.
The ceiling matches the one the Python SDK already applies to that endpoint.

`_current_trace_id()` pads to 32 hex characters like every other trace id in
that file. Roughly one trace in 256 starts with a zero byte, and an unpadded
id for one of those stopped joining its evaluation to its dataset entry.

The async litellm wrapper is covered. Awaiting is what makes translating a
provider rejection possible, and every test drove the blocking wrapper.

Documentation corrections, all verified against the code rather than assumed:

- A comparison weighs every candidate together, and the judge is called twice
  per row by default, not once. The workbench page claimed one call and half
  the cost of swap-and-confirm; `swap_and_reconcile` defaults to true and the
  config form never writes it, so that was wrong there too.
- Candidate shuffling is the default, not the only behaviour.
- The CSV winner column reads `unresolved` only when the judge's answer
  matches none of the candidates it was shown. A winner naming a target the
  run no longer lists is written under its own name.
- `npx langwatch report` scrubs the title, summary and transcript by pattern.
  It does not redact everything: an address passed with `--email` is spliced
  in after the redactor runs, and only IPv4 is scanned.

Also: the comparison test suite is split along behaviour lines and under the
file-size rule, the compatibility spec describes what a person observes
instead of request internals, and the eviction test stopped opening four
thousand real spans to prove one bound.
