+++ docs/evaluations/experiments/sdk.mdx
@@ -494,7 +494,7 @@ Scoring each target on its own leaves you calibrating numbers: is `0.78` better 
 
 ### Comparing a row
 
-Every target that recorded an output for the row is a candidate, and they are all judged in a single call. There is no candidate list to assemble and no judge prompt to choose:
+Every target that recorded an output for the row is a candidate, and one comparison weighs all of them together rather than a round of one-against-one matchups. Behind it the judge is called twice per row by default, the second time with the candidate order reversed, which is what [`swap_and_reconcile`](#options) does; turn that off and one comparison is one judge call. There is no candidate list to assemble and no judge prompt to choose:
 
 <Tabs>
   <Tab title="Python">
@@ -546,9 +546,9 @@ The winner comes back named with the target name you registered. Two candidates 
 </Info>
 
 <Note>
-  `compare()` names the row the way `log()` does in the same SDK. Python is handed the index by `loop()`, so it passes it on as the first argument, and it needs a whole number. TypeScript infers it from the iteration the comparison runs in, so `index` is only needed when you compare outside a `run()` callback, where it also accepts a string row label.
+  `compare()` names the row the way `log()` does in the same SDK. Python is handed the index by `loop()`, so it passes it on as the first argument, and it needs a whole number. TypeScript infers it from the iteration the comparison runs in, so `index` is only needed when you compare outside a `run()` callback, and it is a whole number there too.
 
-  Candidate order is shuffled per row, seeded by the row index, so re-comparing the same row presents the candidates in the same order and a different row shuffles differently. A string label carries no position to seed with, so the judge falls back to its own stable order for it.
+  Candidate order is shuffled per row by default, seeded by the row index, so re-comparing the same row presents the candidates in the same order and a different row shuffles differently. `randomize_order` / `randomizeOrder` governs it: turn it off and the judge sees the candidates in the order their targets were registered.
 </Note>
 
 ### Judging against a reference answer
@@ -690,7 +690,7 @@ With three or more targets there is also a leaderboard that ranks them from ever
 
 ### A complete comparison
 
-Three variants, one judge call per row, and a closing summary that fails a [CI build](/evaluations/experiments/ci-cd) when the run had failures:
+Three variants, one comparison per row, and a closing summary that fails a [CI build](/evaluations/experiments/ci-cd) when the run had failures:
 
 <Tabs>
   <Tab title="Python">
@@ -1009,11 +1009,23 @@ A comparison grades the row as a whole rather than one target, so it gets its ow
 
 | Column Pattern | Description | Example |
 |----------------|-------------|---------|
-| `{comparison}_winner` | Name of the winning target, or `tie` | `comparison_winner` → `gpt-5-mini` |
-| `{comparison}_candidates` | The targets compared on this row | `comparison_candidates` → `gpt-5-mini, claude-sonnet-5` |
+| `{comparison}_winner` | Name of the winning target, `tie`, or empty | `comparison_winner` → `gpt-5-mini` |
+| `{comparison}_candidates` | The targets compared on this row, comma separated | `comparison_candidates` → `gpt-5-mini, claude-sonnet-5` |
 | `{comparison}_reasoning` | The judge's explanation for the verdict | `comparison_reasoning` → `"gpt-5-mini gives the exact reset link"` |
 
-The whole block is empty for a row the judge did not call, and the winner reads `unresolved` on the rare row where the judge answered with a name none of the targets match. A row's candidates can be fewer than the comparison's full set of targets, because a target that produced no output for that row is left out of the matchup.
+What each [verdict](#the-verdict) writes into those three cells:
+
+| Verdict | `{comparison}_winner` | `{comparison}_candidates` | `{comparison}_reasoning` |
+|---------|-----------------------|---------------------------|--------------------------|
+| `decided` | The winning target's name | The targets judged on this row | The judge's explanation |
+| `tie` | `tie` | The targets judged on this row | The judge's explanation |
+| `inconclusive` | Empty | Empty | Empty |
+| `skipped` | Empty | Empty | Empty |
+| `error` | Empty | Empty | Empty |
+
+An empty winner is never a tie. The three cells go empty together for `inconclusive`, `skipped` and `error`, and for a row the comparison never ran on, so `tie` in the winner column always means the judge saw the candidates and called the row even.
+
+A row's candidates can be fewer than the comparison's full set of targets, because a target that produced no output for that row is left out of the matchup. The winner reads `unresolved` on the rare row where the judge's answer matches none of the candidates it was shown, and a winner naming a target the run no longer lists is written under its own name.
 
 ### Example CSV Output
