+++ skills/_compiled/native/setup-lw/SKILL.md
@@ -15,19 +15,18 @@ Get the CLI authenticated and talking to the right LangWatch project, then verif
 
 IMPORTANT: You will need a LangWatch API key. Check if LANGWATCH_API_KEY is already in the project's .env file — most environments already have this provisioned. If they have a LANGWATCH_ENDPOINT in .env, they are on a self-hosted instance, so use that endpoint instead of app.langwatch.ai.
 
-For CI and agents, configure non-interactively — never block on a browser:
+For CI and agents, configure non-interactively, and never block on a browser. Have the runner inject the key from your secret store into `LANGWATCH_API_KEY`, which is what the CLI reads:
 
-```bash
-langwatch login --api-key <key>                          # writes LANGWATCH_API_KEY to .env
-langwatch login --api-key <key> --endpoint https://lw.acme.internal   # self-hosted
+```yaml
+# GitHub Actions; any secret store works the same way
+env:
+  LANGWATCH_API_KEY: ${{ secrets.LANGWATCH_API_KEY }}
+  LANGWATCH_ENDPOINT: https://lw.acme.internal   # self-hosted only; omit for cloud
 ```
 
-Or set the environment directly (what the CLI actually reads):
+Keep the key off the command line: an argument is readable by every process on the machine, and a key typed at a prompt lands in the shell history file.
 
-```bash
-export LANGWATCH_API_KEY=<project-api-key>
-export LANGWATCH_ENDPOINT=https://lw.acme.internal       # self-hosted only; omit for cloud
-```
+Locally, `langwatch login --api-key <key>` is the quick way in. It writes `LANGWATCH_API_KEY` to `.env`, so keep `.env` out of version control. Add `--endpoint https://lw.acme.internal` for a self-hosted instance.
 
 For humans at a terminal, plain `langwatch login` asks interactively (cloud vs self-hosted, AI tools vs project SDK), and `langwatch login --device` runs the RFC 8628 device flow via company SSO.
 
@@ -68,7 +67,7 @@ Discover commands with `langwatch --help` and `langwatch <subcommand> --help`. L
 
 If no shell is available, fetch the same Markdown over plain HTTP. Append `.md` to any docs path (e.g. https://langwatch.ai/docs/integration/python/guide.md). Index: https://langwatch.ai/docs/llms.txt. Scenario index: https://langwatch.ai/scenario/llms.txt
 
-If anything fails or confuses you while following this skill (broken commands, docs that do not match reality, errors you had to work around), ask the user for permission and run `npx langwatch report --user-approved` with a `--title` and `--summary` (or `--session <transcript.jsonl>`) to send it to the LangWatch team. No login needed, secrets and personal data are redacted locally, and it directly shapes what gets fixed. `npx langwatch report --help` explains the options.
+If anything fails or confuses you while following this skill (broken commands, docs that do not match reality, errors you had to work around), ask the user for permission and run `npx langwatch report --user-approved` with a `--title` and `--summary` (or `--session <transcript.jsonl>`) to send it to the LangWatch team, and it directly shapes what gets fixed. No login or API key needed. Nothing is sent without `--user-approved`, and `--dry-run` prints the exact payload without sending anything. The title, summary and transcript are scrubbed locally first, by pattern: secrets and API keys, plus email addresses, phone numbers, card numbers and public IPv4 addresses. Anything no pattern matches is sent as written, including a contact address passed with `--email`, so preview with `--dry-run` when the session touched sensitive data. `npx langwatch report --help` explains the options.
 
 ## Troubleshooting
