+++ skills/recipes/setup-lw/SKILL.mdx
@@ -18,19 +18,18 @@ Get the CLI authenticated and talking to the right LangWatch project, then verif
 
 <ApiKeySetup />
 
-For CI and agents, configure non-interactively — never block on a browser:
+For CI and agents, configure non-interactively, and never block on a browser. Have the runner inject the key from your secret store into `LANGWATCH_API_KEY`, which is what the CLI reads:
 
-```bash
-langwatch login --api-key <key>                          # writes LANGWATCH_API_KEY to .env
-langwatch login --api-key <key> --endpoint https://lw.acme.internal   # self-hosted
+```yaml
+# GitHub Actions; any secret store works the same way
+env:
+  LANGWATCH_API_KEY: ${{ secrets.LANGWATCH_API_KEY }}
+  LANGWATCH_ENDPOINT: https://lw.acme.internal   # self-hosted only; omit for cloud
 ```
 
-Or set the environment directly (what the CLI actually reads):
+Keep the key off the command line: an argument is readable by every process on the machine, and a key typed at a prompt lands in the shell history file.
 
-```bash
-export LANGWATCH_API_KEY=<project-api-key>
-export LANGWATCH_ENDPOINT=https://lw.acme.internal       # self-hosted only; omit for cloud
-```
+Locally, `langwatch login --api-key <key>` is the quick way in. It writes `LANGWATCH_API_KEY` to `.env`, so keep `.env` out of version control. Add `--endpoint https://lw.acme.internal` for a self-hosted instance.
 
 For humans at a terminal, plain `langwatch login` asks interactively (cloud vs self-hosted, AI tools vs project SDK), and `langwatch login --device` runs the RFC 8628 device flow via company SSO.
